修复添加预收费记录默认数据不完整

+++ ruoyi-system/src/main/java/com/ruoyi/system/fantang/domain/FtPrepaymentVo.java
@@ -75,7 +75,7 @@ public class FtPrepaymentVo implements Serializable {
      * 结算标志
      */
     @Excel(name = "结算标志")
-    private Long settlementFlag;
+    private Integer settlementFlag;
 
     /**
      * 预付费金额
